Security fix

+++ traveller/modules/cfp/view.py
@@ -94,46 +94,55 @@ def edit_talk(year, talk_id):
 @module_blueprint.route("/<year>/talk/<talk_id>/rate/<int:score>/<talk_num_>")
 @login_required
 def rate_talk(year, talk_id, score, talk_num_):
-    if score not in [0, 1, 2]:
-        return '---'
-    talk = Talk.query.get(talk_id)
-    reviewers = [sl.reviewer for sl in talk.score_lists]
-    if current_user not in reviewers:
-        
-        score_list = ScoreList()
-        score_list.score = score
-        score_list.reviewer = current_user
-        score_list.talk_id = talk.id
-        talk.score_lists.append(score_list)
-        talk.update()
+    if (current_user.has_role('reviewer') or current_user.is_admin):
+        if score not in [0, 1, 2]:
+            return '---'
+        talk = Talk.query.get(talk_id)
+        reviewers = [sl.reviewer for sl in talk.score_lists]
+        if current_user not in reviewers:
+            
+            score_list = ScoreList()
+            score_list.score = score
+            score_list.reviewer = current_user
+            score_list.talk_id = talk.id
+            talk.score_lists.append(score_list)
+            talk.update()
+        else:
+            for sl in talk.score_lists:
+                if sl.reviewer == current_user:
+                    sl.score = score
+                    sl.update()
+                    break
+        return mhelp.redirect_url('y.review', year=year, talk_num_=talk_num_)
     else:
-        for sl in talk.score_lists:
-            if sl.reviewer == current_user:
-                sl.score = score
-                sl.update()
-                break
-    return mhelp.redirect_url('y.review', year=year, talk_num_=talk_num_)
+        alert_danger('Permission denied!')
+        return mhelp.redirect_url('www.index')
 
 
 @module_blueprint.route("/<year>/talk/<talk_id>/final_talk_action", methods=["GET", "POST"])
 @login_required
 def final_talk_action(year, talk_id):
-    if request.method == 'GET':
-        context = mhelp.context()
-        talk = Talk.query.get(talk_id)
-        AdminTalkForm_ = AdminTalkForm
+    if current_user.is_admin:
+        if request.method == 'GET':
+            context = mhelp.context()
+            talk = Talk.query.get(talk_id)
+            AdminTalkForm_ = AdminTalkForm
 
-        context.update(locals())
-        return render_template('conftheme/{}/parts/final_talk_action.html'.format(year), **context)
-    elif request.method == 'POST':
-        talk = Talk.query.get(talk_id)
-        form = AdminTalkForm(obj=talk)
-        form.populate_obj(talk)
-        form.validate()
+            context.update(locals())
+            return render_template('conftheme/{}/parts/final_talk_action.html'.format(year), **context)
+        elif request.method == 'POST':
+            talk = Talk.query.get(talk_id)
+            form = AdminTalkForm(obj=talk)
+            form.populate_obj(talk)
+            form.validate()
 
-        talk.update()
-        alert_success('Talk status changed!')
-        return mhelp.redirect_url('cfp.final_talk_action', year=year, talk_id=talk_id)
+            talk.update()
+            alert_success('Talk status changed!')
+            return mhelp.redirect_url('cfp.final_talk_action', year=year, talk_id=talk_id)
+
+    else:
+        alert_danger('No permission to view page!')
+        return mhelp.redirect_url('www.index')
 
 
 
@@ -141,5 +150,9 @@ def final_talk_action(year, talk_id):
 @login_required
 def delete_talk(year, talk_id):
     talk = Talk.query.get(talk_id)
-    talk.delete()    
+    if talk.submitter_id == current_user.id:
+        talk.delete()
+        alert_success('Talk deleted!')
+    else:
+        alert_danger("You don't have access to delete talks!")
     return mhelp.redirect_url('y.profile', year=year)
